updated makefile getting segmentation error

+++ debug.c
@@ -4,10 +4,13 @@
  *  Created on: Sep 29, 2017
  *      Author: monish and sanika
  */
+#include<stdint.h>
+#include<stdio.h>
+
 
 void print_memory(uint8_t * start, uint32_t length)
 {
-	short i;
+	uint32_t i;
 	for (i=0;i<length;i++)
 	{
 		printf("%x",*(start+i));
